Add Authentication module for Kubernetes node
This authentication module connects to docker to figure out the pod
name associated to the PID requesting the service by ay of discovering
the container id via the process cgroup namespace.

The pod name as set in the metadata label named 'io.kubernetes.pod.name'
is then used as the 'remote_user' attribute on the request.

Signed-off-by: Simo Sorce <simo@redhat.com>