Improve logging format and configuration
This patch adda an 'origin' argument to the logger formatting that
coms from the configuration parser and ties a log entry to the
implementing class as well as the specific configuration facility
that instantiated it.

Also adds per configuration section debugging unless the global
debug statment is true, in which case all objects have debugging
forcibly turned on.

Signed-off-by: Simo Sorce <simo@redhat.com>
Reviewed-by: Christian Heimes <cheimes@redhat.com>