Refactor CustodiaClient interface
Use the term secret and not key to refer to .. well .. secrets.

Store the last response instead of returning it to the caller, this
way there is a consistent way to get access to it and only as needed.

Change the name to CustodiaSimpleClient in preparaion for extending the
support to other Secret types.

Signed-off-by: Simo Sorce <simo@redhat.com>