Allow to send raw data for the simple type
In the simple type case we can allow clients to send the raw value and
request the raw value by setting the application/octet-stream content
type in the Content-Type header for PUTs and the Accept header for GETs.

The content is arbitrary and will be saved by base64 encoding it on receipt.
If the default 'application/json' type is used to later fetch the secret,
then the value field in the json payload will be a base64 string.

Fixes #10

Signed-off-by: Simo Sorce <simo@redhat.com>