Add sd_notify() call to custodia server
The Custodia server is now able to notify systemd that it is ready to
handle requests. systemd.daemon.notify("READY=1") is called after the
listening socket has been created and before the HTTP server starts its
server loop.

Custodia depends on python-systemd or python3-systemd for notification.
If no systemd bindings are available or systemd isn't running,
sd_notify() is not called.

Signed-off-by: Christian Heimes <cheimes@redhat.com>
Reviewed-by: Simo Sorce <simo@redhat.com>

[Removed stray blank line addition -SSS]

Close #15