Add /etc/custodia as an externally mounted volume in docker image
The /etc/custodia directory should be an externally mounted volume
in our docker image.  Currently, /etc/custodia is a part of the
image itself.  This means that configuration changes will be stored
in the container using the union FS.  Configuration should be mounted
from the host itself, just like we do for the database and logs. This
will allow the container to be upgraded to a new image version without
losing the configuration changes.