Update docs to refer to RTD instead of wiki
The information documented on the Custodia wiki is very minimal.
The only thing described there is the basic plugin architecture.
This patch moves the content from the wiki into the plugin section
of the in-tree docs.  Since these docs are built and published on
RTD, I've also updated our README files to refer to the RTD page
instead of the wiki.  These changes will allow us to get rid of
our wiki and just use RTD as the single official source of online
Custodia documentation.