Fix test flapping due to DNS resolution
Using an actual DNS name sometimes causes the test to fail with error
dependent on resolution.
Given this test is meant to check for connection errors and not for name
resolution errors, change the name to use localhost and an unassigned
low TCP port so we get always a connection refused error.

Signed-off-by: Simo Sorce <simo@redhat.com>
Reviewed-by: Raildo Mascena <rmascena@redhat.com>
Closes: #143