Distiguish between containes and keys
When we request the contents of a cointainer, we return the names of
keys and sub-containers as plain names, this makes it impossible to
know which of them refer to actual keys and which refer to
sub-containers.

When listing container content we now add a '/' suffix to the names of
sub-containers, so they can be easily distinguished from keys.

Signed-off-by: Raildo Mascena <rmascena@redhat.com>
Reviewed-by: Simo Sorce <simo@redhat.com>
Resolves: #132
Closes: #149