Plugin config: add special marker for inherit global
Sometimes it is desirable to inherit some values from [global] block,
e.g. 'debug' or 'tls_cafile'. Instead of a special case just for
'debug', I rather implemented a special value.

Example:

class ExamplePlugin(CustodiaPlugin)
    debug = PluginOption(bool, INHERIT_GLOBAL(False), '')

The value for 'debug' is first looked up in the config block for the
plugin. If the config block has no 'debug' stanza, then it is looked up
in [global]. In case global does have the setting either, it is set to
False.

Closes: #166
Signed-off-by: Christian Heimes <cheimes@redhat.com>