Add secret_pinning option to encgen driver
This option adds the key name into the protected header of the JWE token
used to encryp secrets. This allows Custodia to verify that the database
was not tampered with (e.g. secrets swapped between keys).

If enabled this option will cause exception when pre-exsiting keys
generated without the protected header are looked up.
however turning off this option will not cause failures.

Signed-off-by: Simo Sorce <simo@redhat.com>