Remove etcd storage plugin
The etcd storage plugin was added as a tech demo and proof of concept
to show, how an external database could be used to store secrets. It
hasn't been evolved into a production-ready storage plugin. Further more
it is limited to etcd2 API and not etcd3's new gPRC.

Since it's not usable and causes additional overhead in testing and
downstream packaging, we decided to remove the plugin from the core
distribution.

Signed-off-by: Christian Heimes <cheimes@redhat.com>